feature/refactor how to import agents
